# openapi.yaml
openapi: 3.0.1
info:
  title: Code Writer Plugin
  description: A plugin that allows the user to write code to a specific directory on their local hard drive. If the directory doesn't exist, it will be created. If no directory is specified, a temporary directory with a random number will be created and the new file will be written into that directory.
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /write_code:
    post:
      operationId: writeCode
      summary: Write code to a specific directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/writeCodeRequest'
      responses:
        "200":
          description: OK

  /read_file:
    post:
      operationId: readFile
      summary: Read a file from a specific path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/readFileRequest'
      responses:
        "200":
          description: OK

  /run_python_code:
    post:
      operationId: runPythonCode
      summary: Run Python code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/runPythonCodeRequest'
      responses:
        "200":
          description: OK

  /run_python_ast_code:
    post:
      operationId: runPythonAstCode
      summary: Run Python code with AST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/runPythonAstCodeRequest'
      responses:
        "200":
          description: OK

components:
  schemas:
    writeCodeRequest:
      type: object
      required:
      - code
      properties:
        code:
          type: string
          description: The code to write.
          required: true
        directory:
          type: string
          description: The directory to write to.
        filename:
          type: string
          description: The filename for the new file.

    readFileRequest:
        type: object
        required:
        - filepath
        properties:
          filepath:
            type: string
            description: The path of the file to read.
            required: true

    runPythonCodeRequest:
      type: object
      required:
      - code
      properties:
        code:
          type: string
          description: The Python code to run.

    runPythonAstCodeRequest:
      type: object
      required:
      - code
      properties:
        code:
          type: string
          description: The Python code to run with AST.